added reset command and button on the gui so users can reset their scale to the default value and updated the languange files so the key category would be set correctly

+++ src/client/java/frank1o3/statscale/client/ScaleClientState.java
@@ -89,7 +89,7 @@ public final class ScaleClientState {
      *
      * @param scale The scale value just sent to the server.
      */
-    public static void setCurrentScale(float scale) {
+    public static void setCurrentScale(double scale) {
         currentScale = scale;
     }
